Refactoring to move API calls to App.jsx
