+++ server/index.js
@@ -19,9 +19,9 @@ app.get('/products/:id/styles', product.styles);
 app.post('/cart', product.cart);
 
 //Reviews
-app.get('/reviews', reviews.getReviews);
-app.post('/sortReviews', reviews.getReviews);
-app.get('/reviewsMeta', reviews.getReviewMeta);
+app.get('/reviews/:pid', reviews.getReviews);
+app.post('/sortReviews/:pid', reviews.getReviews);
+app.get('/reviewsMeta/:pid', reviews.getReviewMeta);
 app.post('/reviews', reviews.postReview);
 app.post('/reviews/:review_id/helpful', reviews.markHelpfulReview);
